task2_chatbots/dfbackend/requirements.txt
alembic==0.9.5
bcrypt==3.1.3
cffi==1.10.0
click==6.7
coverage==4.4.1
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
ForgeryPy==0.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
nose==1.3.7
psycopg2==2.7.3
pycparser==2.18
PyJWT==1.5.2
python-dateutil==2.6.1
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
dialogflow==0.6.0 
pusher==2.1.3
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
oauth2client==4.1.3 
